+++ RELEASE-NOTES.md
@@ -8,6 +8,7 @@
 * #122 Expression API
 * #123 Improve performance of "DF.filterRows(BooleanSeries)" 
 * #125 Column-based sort algorithm for better performance and expressions support
+* #127 Make Aggregator compatible with SeriesExp
 
 ## Release 0.10